pyproject.toml
[tool.poetry]
name = "django-template"
version = "0.1.0"
description = ""
authors = ["myounghumbaek <myounghum.baek@heumlabs.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
django-environ = "^0.11.2"
django-ninja = "^1.3.0"


[tool.poetry.group.test.dependencies]
jupyter = "^1.0.0"
pytest = "^8.3.2"
flake8 = "^7.1.1"
pyflakes = "^3.2.0"
coverage = "^7.6.1"
pycodestyle = "^2.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line_length = 119
exclude = '/(\.venv|venv|\.env|env|migrations)/'

[tool.isort]
line_length = 119
profile = "black"
skip_glob = "^(manage.py|migrations)"
src_paths = ["app"]

[tool.flake8]
ignore = "I101,W503,B008,B019,E501"
exclude = ".git,*/migrations/*,settings.py,manage.py,settings_*.py,*/query.py"
max-line-length = 119

